scripts/requirements.txt
coloredlogs==15.0.1
dobotapi==0.0.3
future==1.0.0
humanfriendly==10.0
iso8601==2.1.0
paho-mqtt==2.1.0
pip==26.0.1
prompt-toolkit==3.0.52
pyreadline3==3.5.4
pyserial==3.5
pyyaml==6.0.3
serial==0.0.97
svg-path==6.3
wcwidth==0.6.0